% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINARY_DF.R, R/CONT_DF.R
\docType{data}
\name{BINARY_DF}
\alias{BINARY_DF}
\title{Toy data with binary outcome}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
#' \describe{
 \item{ID}{ID's of individuals}
 \item{SEX}{sex of individuals (character)}
 \item{AGE}{age (numeric)}
 \item{COV_Q_OTHER}{quantitative covariates (numeric)}
 \item{COV_C_OTHER}{categorical covariate (categorical)}
 \item{PC_6-14}{genetic principal components (randomly generated)}
 \item{VAR1-18}{numeric exposure variables (numeric)}
 \item{VAR19-35}{categorical exposure variables (categorical)}
 \item{PHENO}{binary phenotype (0/1)}
 \item{TIME}{survival time (months)}
}

An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
#' \describe{
 \item{ID}{ID's of individuals}
 \item{SEX}{sex of individuals (character)}
 \item{AGE}{age (numeric)}
 \item{COV_Q_OTHER}{quantitative covariates (numeric)}
 \item{COV_C_OTHER}{categorical covariate (categorical)}
 \item{PC_6-14}{genetic principal components (randomly generated)}
 \item{VAR1-18}{numeric exposure variables (numeric)}
 \item{VAR19-35}{categorical exposure variables (categorical)}
 \item{PHENO}{continuous phenotype (numeric)}
}
}
\usage{
data(BINARY_DF)

data(CONT_DF)
}
\description{
Artificially generated toy data set of 5000 individuals with a binary outcome
 and a set of categorical and continuous exposure variables

Artificially generated toy data set of 5000 individuals with a numeric outcome
 and a set of categorical and continuous exposure variables
}
\examples{

data(BINARY_DF)
head(BINARY_DF)


data(BINARY_DF)
head(BINARY_DF)

}
\references{
This data set was artificially created for the PXStools package.

This data set was artificially created for the PXStools package.
}
\keyword{datasets}
