% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_xwas.R
\name{manhattan_xwas}
\alias{manhattan_xwas}
\title{manhattan_xwas}
\usage{
manhattan_xwas(xdff, pval, thresh = 0.05)
}
\arguments{
\item{xdff}{matrix returned from XWAS function, rownames of matrix should be the X variables}

\item{pval}{column name of p value}

\item{thresh}{p value threshold for signficance}
}
\description{
this is a function plot the p values of the XWAS results, analagous to a manhattan plot.
  since the y axis is in the -log scale, there may be issues with plotting if the p value is zero or very close to zero (taking the neg log of it will be infinite)
}
\author{
Yixuan He, \email{yixuan_he@hms.harvard.edu}
}
