% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coeff_xwas.R
\name{plot_coeff_xwas}
\alias{plot_coeff_xwas}
\title{plot_coefficients}
\usage{
plot_coeff_xwas(xdff, pval, coeff, thresh = 0.05, all = FALSE)
}
\arguments{
\item{xdff}{matrix returned from XWAS function, rownames of matrix should be the X variables}

\item{pval}{column name of p value}

\item{coeff}{column name of coefficients}

\item{thresh}{p value threshold for signficance}

\item{all}{default is to plot only signficant associaitons, all=TRUE plots all associatons}
}
\description{
this is a function plot the coefficients XWAS results
}
\author{
Yixuan He, \email{yixuan_he@hms.harvard.edu}
}
