% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PXS.R
\name{PXS}
\alias{PXS}
\title{PXS}
\source{

}
\usage{
PXS(df, X, cov, mod, IDA, IDB, IDC, seed, removes = NULL, intermediate = F)
}
\arguments{
\item{df}{the data frame inpt}

\item{X}{column name of signficant exposure variables from XWAS}

\item{cov}{column name of covariates}

\item{mod}{type of model to run; 'lm' for linear regression, 'logistic' for logistic regression; 'cox' for Cox regression}

\item{IDA}{list of IDs to from XWAS procedure}

\item{IDB}{list of IDs for testing set}

\item{IDC}{list of IDs in the final prediction set}

\item{seed}{setting a seed}

\item{removes}{any exposure response to remove from XWAS, in the form of a list}

\item{intermediate}{whether or not to save intermediate files}

\item{fdr}{whether or not to adjust for multiple hypothesis correctin}
}
\description{
this function builds a polyexposure risk score
}
\author{
Yixuan He, \email{yixuan_he@hms.harvard.edu}
}
