% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwas.R
\name{xwas}
\alias{xwas}
\title{xwas}
\usage{
xwas(df, X, cov, mod, IDA, removes = NULL, fdr = TRUE)
}
\arguments{
\item{df}{the data frame inpt}

\item{X}{column name of exposure variables to run XWAS}

\item{cov}{column name of covariates}

\item{mod}{type of model to run; 'lm' for linear regression, 'logistic' for logistic regression; 'cox' for Cox regression}

\item{IDA}{list of IDs to include in XWAS}

\item{removes}{any exposure response to remove from XWAS, in the form of a list}

\item{fdr}{whether or not to adjust for multiple hypothesis correctin}
}
\description{
this is a function to run an exposure wide association study (XWAS) with any phenotype and a set of exposure variables
}
\author{
Yixuan He, \email{yixuan_he@hms.harvard.edu}
}
